[vichan] add generic extractors for vichan imageboards
includes 8kun.top, smuglo.li, and wikieat.club
