+++ gallery_dl/extractor/__init__.py
@@ -17,7 +17,6 @@ modules = [
     "4chan",
     "500px",
     "8chan",
-    "8kun",
     "8muses",
     "adultempire",
     "architizer",
@@ -142,6 +141,7 @@ modules = [
     "twitter",
     "unsplash",
     "vanillarock",
+    "vichan",
     "vk",
     "vsco",
     "wallhaven",
@@ -151,7 +151,6 @@ modules = [
     "webtoons",
     "weibo",
     "wikiart",
-    "wikieat",
     "xhamster",
     "xvideos",
     "zerochan",
@@ -161,7 +160,6 @@ modules = [
     "foolslide",
     "mastodon",
     "shopify",
-    "smugloli",
     "lolisafe",
     "imagehosts",
     "directlink",
